+++ src/Support/WebBlocks.php
@@ -10,7 +10,7 @@ final class WebBlocks
 
   public const HANDLE = 'webblocks-cms';
 
-  public const VERSION = '1.40.8';
+  public const VERSION = '1.40.9';
 
   public const UI_VERSION = 'v2.10.0';